requirements.txt
streamlit>=1.39
streamlit-mic-recorder>=0.0.8
openai>=1.54
duckdb>=1.1
chromadb>=0.5
pandas>=2.2
python-dotenv>=1.0
numpy>=1.26
